add map-reduce operation
Add map-reduce operation to find top and bottom apps that use one or
more permission in a list of permissions and sort the results by
download count. The results can be limited to a number of docs.
The new command to run this map-reduce task is called
find_top_bottom_apps_in_any_permission